+++ game/src/util/random.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 void init_random();
-float uniform_random(float min, float max);
-int discrete_random(int min, int max);
+double uniform_random(double min, double max);
+int32_t discrete_random(int32_t min, int32_t max);
